Improved scripts
Improved wording and explanation on the essential tweaks and essential undo scripts so that they restore and remove the same things in case a user want to revert back to the old configuration.

Also added GodMode as some users wanted that to be included here.

TODO:
Add the 5 last placement holder items.