Customize About this PC
Added a new section so that users of this script can update information about this PC, it should detect if a key exist so it can be used on Windows 10 too but the main goal is to achieve this on Windows 11. Also implemented so that the user can set a default background image.